+++ apps/web/app/(pages)/creators/page.tsx
@@ -5,7 +5,7 @@ import Footer from "@/components/Layout/Footer";
 import { PageContainer, Main } from "../../styles";
 import CreatorsSection from "./index";
 import ShortStory from "./creators/ShortStory";
-import WhyCreatorsChoose from "./creators/WhyCreatorsChoose";
+import WhyCreatorsChoose from "../../../components/Feature/ExploreCreators/Creators/WhyCreatorsChoose";
 
 export default function CreatorsPage() {
   return (
